fix: move memo helper to utils and use shorthand hook names
